requirements.txt
polars
black
environs

